feat: 사용자 이름이 비어 있으면 예외를 발생시킨다
