+++ src/test/java/domain/player/PlayersTest.java
@@ -1,5 +1,9 @@
 package domain.player;
 
+import domain.player.Name;
+import domain.player.Participant;
+import domain.player.Player;
+import domain.player.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +20,7 @@ class PlayersTest {
         final List<Player> players = new ArrayList<>();
 
         for (int i = 0; i < 9; i++) {
-            players.add(new Player(new Name("teba" + i)));
+            players.add(new Participant(new Name("teba" + i)));
         }
 
         assertThatCode(() -> new Players(players)).isInstanceOf(IllegalArgumentException.class);
@@ -26,18 +30,18 @@ class PlayersTest {
     @DisplayName("플레이어의 이름이 겹치면 예외가 발생한다")
     void duplicateName() {
         final List<Player> players = new ArrayList<>();
-
-        players.add(new Player(new Name("teba")));
-        players.add(new Player(new Name("teba")));
+        players.add(new Participant(new Name("teba")));
+        players.add(new Participant(new Name("teba")));
 
         assertThatCode(() -> new Players(players)).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
-    @DisplayName("플레이어의 이름을 찾을 때 존재하지 않는 이름이라면 예외가 발생한다")
-    void nameNotExist() {
-        final Players players = Players.fromNames(List.of("a", "b", "c"));
+    @DisplayName("플레이어의 이름 딜러라면 예외가 발생한다")
+    void dealerName() {
+        final List<Player> players = new ArrayList<>();
+        players.add(new Participant(new Name("딜러")));
 
-        assertThatCode(() -> players.findPlayerByName("d")).isInstanceOf(IllegalArgumentException.class);
+        assertThatCode(() -> new Players(players)).isInstanceOf(IllegalArgumentException.class);
     }
 }
